requirements.txt
Django>=6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
django-filter>=25.1
drf-spectacular>=0.29
django-environ>=0.12
mysqlclient>=2.2.1
